requirements.txt
Flask
psycopg2-binary
SQLAlchemy
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
Werkzeug


Flask-Login


Gunicorn
